More class splitting for codefactor.
