+++ Blish HUD/Common/UI/Views/IControlView.cs
@@ -1,0 +1,11 @@
+﻿using Blish_HUD.Controls;
+
+namespace Blish_HUD.Common.UI.Views {
+
+    public interface IControlView<out TControl> where TControl : Control {
+
+        TControl Control { get; }
+
+    }
+
+}
